pyproject.toml
[project]
name = "transformers-course"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
  "transformers[sentencepiece]==4.57.6",
  "datasets==4.5.0",
  "torch==2.9.1",
  "pillow == 12.0.0",
  "evaluate == 0.4.6",
  "accelerate >= 0.35.0",
  "numpy<2.4.0",
  "scipy>=1.17.0",
  "scikit-learn>=1.8.0",
  "wandb==0.24.0",
  "requests>=2.32.5",
  "ipykernel>=7.1.0",
  "pandas-stubs~=2.3.3",
  "pandas>=2.3.3",
  "faiss-cpu>=1.13.2",
  "seqeval>=1.2.2",
  "label-studio>=1.22.0",
  "doccano[postgresql]>=1.5.5",
  "gradio>=6.5.1",
  "gradio-client>=2.0.3",
  "tensorflow>=2.20.0",
  "tf-keras>=2.20.1",
  "huggingface-hub>=0.36.0",
  "trl>=0.29.0",
  "peft>=0.18.1",
]

[tool.black]
# Maksymalna długość linii. 88 to domyślny standard Blacka.
# Wiele zespołów zmienia to na 100 lub 120, jeśli kod jest mocno zagnieżdżony.
line-length = 120

# Określa, dla jakich wersji Pythona ma być generowany kod.
# Pomaga Blackowi zdecydować, czy używać np. nowoczesnych udogodnień składniowych.
target-version = ['py310', 'py311', 'py312']

# Czy Black ma zamieniać pojedyncze cudzysłowy (') na podwójne (")?
# Black domyślnie wymusza ", ustawienie 'false' pozwala mu ich nie zmieniać.
skip-string-normalization = false

# Jeśli używasz Blacka razem z narzędziem 'isort', to ustawienie
# dba o to, by formatowanie magii przecinków było spójne.
preview = false

# Lista plików lub folderów, które Black ma ignorować.
force-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[dependency-groups]
dev = [
    "black>=26.1.0",
]
